minor bugfixes
git-svn-id: http://datajoint.googlecode.com/svn/trunk/matlab@473 2f38a6b3-5502-11bb-9305-b9e7ad8f4d38

+++ erdpdf.m
@@ -5,5 +5,11 @@ erd(entity)
 
 set(gcf,'PaperUnits','inches','PaperPosition',[0 0 11 8],'PaperSize',[11 8])
 print('-dpdf', './temp_erd')
-system('open ./temp_erd.pdf; rm ./temp_erd.pdf');
+
+if ~strncmpi(computer, 'MAC', 3)
+    disp 'Saved temp erd'
+else
+    % on a mac, open the PDF and erase the file
+    system('open ./temp_erd.pdf; rm ./temp_erd.pdf');
+end
 close(gcf)
